requirements.txt
pandas==0.24.2
#  opencv-python==4.1.0.25
tensorflow==1.12.0
scikit-learn==0.20.3
xlrd==1.2.0
pymysql==0.9.3
Pillow==6.0.0
flask==1.0.2
requests==2.21.0
gevent==1.4.0